refactoring test code into multiple files
